fix(nexus-web): keep the workspace shell mounted across switches
Auth and the chrome lived under [workspaceId], so changing workspace
remounted the whole shell: full-screen Loading, GET /api/auth/me, then
a fetch waterfall. Lift that layout one level up and only sync the
URL workspace id into the store.
