+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/layout.tsx
@@ -1,133 +1,11 @@
 'use client';
 
-import { Component, useEffect, useState } from 'react';
+import { useEffect, useLayoutEffect } from 'react';
 import { useParams, usePathname, useRouter } from 'next/navigation';
-import { isWorkspacePathAllowed } from '@/lib/feature-access';
-import { WorkspaceLayout } from '@/components/shell/workspace-layout';
-import { ShellTitleProvider } from '@/components/shell/shell-title';
+import { getWorkspaceSwitchPath, isWorkspacePathAllowed, pathNeedsAgentCatalog, pathNeedsGraphExport } from '@/lib/feature-access';
 import { GraphExportToastHost } from '@/components/graph/graph-export-toast-host';
 import { useWorkspaceStore } from '@/stores/workspace';
-import { clearAuthFlagCookie } from '@/lib/auth-session';
 import { useAuthStore } from '@/stores/auth';
-
-// Catches unhandled promise rejections (not caught by React error boundaries)
-// Next.js uses thrown sentinels (NEXT_REDIRECT, NEXT_NOT_FOUND) to drive
-// navigation from Server Components. They are not real errors and must not
-// surface in any user-facing error UI.
-function isNextNavigationSentinel(reason: unknown): boolean {
-  if (!reason) return false;
-  const digest = (reason as { digest?: unknown }).digest;
-  if (typeof digest === 'string' && digest.startsWith('NEXT_')) return true;
-  const message = (reason as { message?: unknown }).message;
-  if (typeof message === 'string' && (message === 'NEXT_REDIRECT' || message === 'NEXT_NOT_FOUND')) {
-    return true;
-  }
-  return false;
-}
-
-function AsyncErrorCatcher() {
-  const [asyncError, setAsyncError] = useState<string | null>(null);
-
-  useEffect(() => {
-    const handleUnhandledRejection = (event: PromiseRejectionEvent) => {
-      if (isNextNavigationSentinel(event.reason)) {
-        event.preventDefault();
-        return;
-      }
-      const msg = event.reason?.message || event.reason?.toString() || 'Unknown async error';
-      const stack = event.reason?.stack || '';
-      console.error('[NEXUS Unhandled Rejection]', msg, stack);
-      setAsyncError(`${msg}\n\n${stack}`);
-      event.preventDefault();
-    };
-
-    const handleError = (event: ErrorEvent) => {
-      if (isNextNavigationSentinel(event.error) || event.message === 'Uncaught Error: NEXT_REDIRECT' || event.message === 'Uncaught Error: NEXT_NOT_FOUND') {
-        return;
-      }
-      console.error('[NEXUS Global Error]', event.message, event.error?.stack);
-      setAsyncError(`${event.message}\n\n${event.error?.stack || ''}`);
-    };
-
-    window.addEventListener('unhandledrejection', handleUnhandledRejection);
-    window.addEventListener('error', handleError);
-    return () => {
-      window.removeEventListener('unhandledrejection', handleUnhandledRejection);
-      window.removeEventListener('error', handleError);
-    };
-  }, []);
-
-  if (!asyncError) return null;
-
-  return (
-    <div style={{ position: 'fixed', bottom: 16, right: 16, zIndex: 99999, maxWidth: 500, backgroundColor: '#1a1a2e', border: '2px solid #e94560', borderRadius: 8, padding: 16, boxShadow: '0 4px 20px rgba(0,0,0,0.5)' }}>
-      <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: 8 }}>
-        <span style={{ color: '#ff6b6b', fontWeight: 'bold', fontSize: 14 }}>Async Error Caught</span>
-        <button onClick={() => setAsyncError(null)} style={{ color: '#a8a8a8', background: 'none', border: 'none', fontSize: 18, cursor: 'pointer' }}>x</button>
-      </div>
-      <pre style={{ color: '#a8a8a8', fontSize: 11, whiteSpace: 'pre-wrap', wordBreak: 'break-all', maxHeight: 200, overflow: 'auto', margin: 0 }}>{asyncError}</pre>
-    </div>
-  );
-}
-
-// Error boundary to catch React rendering crashes and show a useful message
-class WorkspaceErrorBoundary extends Component<
-  { children: React.ReactNode },
-  { hasError: boolean; error: Error | null }
-> {
-  constructor(props: { children: React.ReactNode }) {
-    super(props);
-    this.state = { hasError: false, error: null };
-  }
-
-  static getDerivedStateFromError(error: Error) {
-    return { hasError: true, error };
-  }
-
-  componentDidCatch(error: Error, info: React.ErrorInfo) {
-    console.error('[WorkspaceErrorBoundary]', error, info.componentStack);
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return (
-        <>
-          <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '100vh', width: '100vw', backgroundColor: '#1a1a2e', padding: 32 }}>
-            <div style={{ maxWidth: 600, textAlign: 'center' }}>
-              <h2 style={{ color: '#ff6b6b', fontSize: 24, fontWeight: 'bold', marginBottom: 16 }}>NEXUS Crash Report</h2>
-              <div style={{ backgroundColor: '#16213e', border: '1px solid #e94560', borderRadius: 8, padding: 16, textAlign: 'left', marginBottom: 16, maxHeight: 300, overflow: 'auto' }}>
-                <p style={{ color: '#ff6b6b', fontSize: 14, fontWeight: 'bold', marginBottom: 8 }}>{this.state.error?.message}</p>
-                <pre style={{ color: '#a8a8a8', fontSize: 11, whiteSpace: 'pre-wrap', wordBreak: 'break-all' }}>
-                  {this.state.error?.stack}
-                </pre>
-              </div>
-              <div style={{ display: 'flex', gap: 8, justifyContent: 'center' }}>
-                <button
-                  onClick={() => this.setState({ hasError: false, error: null })}
-                  style={{ backgroundColor: '#e94560', color: 'white', border: 'none', borderRadius: 6, padding: '8px 20px', fontSize: 14, cursor: 'pointer' }}
-                >
-                  Try again
-                </button>
-                <button
-                  onClick={() => {
-                    localStorage.removeItem('nexus-workspace-storage');
-                    localStorage.removeItem('nexus-auth-storage');
-                    window.location.href = '/auth/login';
-                  }}
-                  style={{ backgroundColor: '#16213e', color: '#e94560', border: '1px solid #e94560', borderRadius: 6, padding: '8px 20px', fontSize: 14, cursor: 'pointer' }}
-                >
-                  Clear cache &amp; login
-                </button>
-              </div>
-            </div>
-          </div>
-          <div hidden>{this.props.children}</div>
-        </>
-      );
-    }
-    return this.props.children;
-  }
-}
 
 export default function WorkspaceIdLayout({
   children,
@@ -138,97 +16,59 @@ export default function WorkspaceIdLayout({
   const pathname = usePathname();
   const router = useRouter();
   const workspaceId = params.workspaceId as string;
-  
-  const { workspaces, currentWorkspaceId, setCurrentWorkspace, syncWorkspaceConversations, fetchWorkspaces } = useWorkspaceStore();
-  const { token } = useAuthStore();
-  const [authReady, setAuthReady] = useState(false);
-  const [membershipChecked, setMembershipChecked] = useState(false);
+
+  const workspaces = useWorkspaceStore((s) => s.workspaces);
+  const currentWorkspaceId = useWorkspaceStore((s) => s.currentWorkspaceId);
+  const setCurrentWorkspace = useWorkspaceStore((s) => s.setCurrentWorkspace);
+  const syncWorkspaceConversations = useWorkspaceStore((s) => s.syncWorkspaceConversations);
+  const contextPanelOpen = useWorkspaceStore((s) => s.contextPanelOpen);
+  const token = useAuthStore((s) => s.token);
   const currentWorkspace = workspaces.find((w) => w.id === workspaceId);
 
-  // Wait for the auth store to rehydrate from localStorage, then silently
-  // validate / refresh the session. authReady becomes true only after
-  // checkAuth() resolves so the redirect (authReady && !token) never fires
-  // prematurely while a token refresh is in flight.
-  useEffect(() => {
-    let cancelled = false;
-    let started = false;
-
-    const runAuth = async () => {
-      if (started) return;
-      started = true;
-      await useAuthStore.getState().checkAuth();
-      if (!cancelled) setAuthReady(true);
-    };
-
-    const unsub = useAuthStore.persist.onFinishHydration(() => {
-      void runAuth();
-    });
-
-    // If already hydrated (e.g., navigating between pages in the same session)
-    if (useAuthStore.persist.hasHydrated()) {
-      void runAuth();
+  // Before paint, so section pages fetch the target workspace instead of
+  // rendering one frame against the previous id.
+  useLayoutEffect(() => {
+    if (workspaceId && workspaceId !== currentWorkspaceId && currentWorkspace) {
+      setCurrentWorkspace(workspaceId);
     }
+  }, [workspaceId, currentWorkspaceId, currentWorkspace, setCurrentWorkspace]);
 
-    return () => {
-      cancelled = true;
-      unsub();
-    };
-  }, []);
-
-  // Redirect to login if not authenticated (after hydration)
   useEffect(() => {
-    if (authReady && !token) {
-      // Drop stale middleware cookie so /auth/login is not bounced back here.
-      clearAuthFlagCookie();
-      router.replace(`/auth/login?redirect=/workspace/${workspaceId}/chat`);
-    }
-  }, [authReady, token, router, workspaceId]);
+    if (!workspaceId) return;
+    if (!contextPanelOpen && !pathNeedsAgentCatalog(pathname)) return;
+    void syncWorkspaceConversations(workspaceId);
+  }, [workspaceId, pathname, contextPanelOpen, syncWorkspaceConversations]);
 
-  // After auth, confirm the user still has at least one workspace membership.
+  // Unknown id: wait for a fresh list before bouncing. A stale in-memory
+  // list used to send people to workspaces[0] on the first paint of a switch,
+  // which cancelled navigation to a workspace that was already in the dropdown.
   useEffect(() => {
-    if (!authReady || !token) return;
+    if (!workspaceId) return;
+    if (workspaces.some((w) => w.id === workspaceId)) return;
+
     let cancelled = false;
-    void fetchWorkspaces().finally(() => {
-      if (!cancelled) setMembershipChecked(true);
+    void useWorkspaceStore.getState().fetchWorkspaces().then(() => {
+      if (cancelled) return;
+      const list = useWorkspaceStore.getState().workspaces;
+      if (list.some((w) => w.id === workspaceId)) return;
+      if (list.length === 0) return;
+      const fallback = list[0];
+      router.replace(
+        getWorkspaceSwitchPath({
+          pathname: pathname || '',
+          targetWorkspaceId: fallback.id,
+          role: fallback.currentUserRole,
+          workspaceFlags: fallback.featureFlags,
+        }),
+      );
     });
     return () => {
       cancelled = true;
     };
-  }, [authReady, token, fetchWorkspaces]);
+  }, [workspaceId, workspaces, pathname, router]);
 
   useEffect(() => {
-    if (!membershipChecked || !token) return;
-    if (workspaces.length === 0) {
-      router.replace('/no-workspace');
-    }
-  }, [membershipChecked, token, workspaces.length, router]);
-
-  // Sync URL workspaceId with store
-  useEffect(() => {
-    if (!membershipChecked) return;
-    if (workspaceId && workspaceId !== currentWorkspaceId) {
-      // Check if workspace exists
-      const workspace = workspaces.find(w => w.id === workspaceId);
-      if (workspace) {
-        setCurrentWorkspace(workspaceId);
-        void syncWorkspaceConversations(workspaceId);
-      } else if (workspaces.length > 0) {
-        // Redirect to first workspace if not found
-        router.replace(`/workspace/${workspaces[0].id}/chat`);
-      } else {
-        router.replace('/no-workspace');
-      }
-    }
-  }, [membershipChecked, workspaceId, currentWorkspaceId, workspaces, setCurrentWorkspace, syncWorkspaceConversations, router]);
-
-  useEffect(() => {
-    if (workspaceId) {
-      void syncWorkspaceConversations(workspaceId);
-    }
-  }, [workspaceId, syncWorkspaceConversations]);
-
-  useEffect(() => {
-    if (!authReady || !token || !workspaceId || !pathname || !currentWorkspace) {
+    if (!token || !workspaceId || !pathname || !currentWorkspace) {
       return;
     }
 
@@ -246,29 +86,12 @@ export default function WorkspaceIdLayout({
     if (notAvailablePath !== pathname) {
       router.replace(notAvailablePath);
     }
-  }, [authReady, token, workspaceId, pathname, currentWorkspace, router]);
-
-  const showShell = authReady && !!token;
+  }, [token, workspaceId, pathname, currentWorkspace, router]);
 
   return (
-    <WorkspaceErrorBoundary>
-      <AsyncErrorCatcher />
-      {/* `children` must appear in exactly ONE place in this tree. It used to be
-          rendered here as well as inside the shell below, and because the two
-          positions do not reconcile, React unmounted and remounted the entire
-          page subtree the moment `showShell` flipped — every route fired all of
-          its data fetches twice per load, the first time before auth was even
-          ready. Render the spinner alone while the session is being checked. */}
-      {!showShell ? (
-        <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', height: '100vh', backgroundColor: '#0a0a1a' }}>
-          <div style={{ color: '#666', fontSize: 14 }}>Loading...</div>
-        </div>
-      ) : (
-        <ShellTitleProvider>
-          <GraphExportToastHost workspaceId={workspaceId} />
-          <WorkspaceLayout>{children}</WorkspaceLayout>
-        </ShellTitleProvider>
-      )}
-    </WorkspaceErrorBoundary>
+    <>
+      {pathNeedsGraphExport(pathname) && <GraphExportToastHost workspaceId={workspaceId} />}
+      {children}
+    </>
   );
 }
